Rework fifo and socket routines
